parser/requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
exceptiongroup==1.0.0
fake-user-agent==2.1.7
fake-useragent==0.1.14
frozenlist==1.3.1
h11==0.14.0
idna==3.4
lxml==4.9.1
multidict==6.0.2
outcome==1.2.0
pip==21.3.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2022.6
requests==2.28.1
selenium==4.5.0
setuptools==60.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio-websocket==0.9.2
trio==0.22.0
urllib3==1.26.12
useragent==0.1.1
wheel==0.37.1
wsproto==1.2.0
yarl==1.8.1
